Ability Score Todo Items - add todo items for ability score choices - change priority to use doubles to allow for easier todo item insertion - convert AbilityScoreManager.getAbilityScoreChoices to an observable list - create property predicate for observing a predicate on a property

+++ src/ui/todo/Priority.java
@@ -5,17 +5,17 @@ import java.util.Arrays;
 import java.util.List;
 
 public class Priority implements Comparable<Priority> {
-    private final List<Integer> priority;
+    private final List<Double> priority;
 
     /**
      * Create a priority
      * @param priority A string in the format "\\d+(\\.\\d)*" (e.g. 13.1.2)
      */
-    public Priority(Integer... priority) {
+    public Priority(Double... priority) {
         this.priority = new ArrayList<>(Arrays.asList(priority));
     }
 
-    public void append(int subPriority) {
+    public void append(double subPriority) {
         priority.add(subPriority);
     }
 
@@ -24,14 +24,14 @@ public class Priority implements Comparable<Priority> {
         int maxSize = Math.max(priority.size(), o.priority.size());
         int result = 0;
         for(int i = 0; i < maxSize; i++) {
-            result = Integer.compare(getOrDefault(priority, i), getOrDefault(o.priority, i));
+            result = Double.compare(getOrDefault(priority, i), getOrDefault(o.priority, i));
             if(result != 0)
                 return result;
         }
         return result;
     }
 
-    private int getOrDefault(List<Integer> list, int index) {
+    private double getOrDefault(List<Double> list, int index) {
         return (index < list.size()) ? list.get(index) : 0;
     }
 }
